Refactor Database code to avoid overload
